Block counting algorithm good enough to test

+++ CMORE/FrameProcessing/BlockDetector.swift
@@ -55,6 +55,6 @@ struct BlockDetector {
 }
 
 // block length 2.5cm or 1in
-func blockLengthInPixels(scale: Float) -> Float {
+func blockLengthInPixels(scale: Double) -> Double {
     return 2.5 / scale
 }
